update tests to get non-zero values; move fixtures to fixtures.py
